Extract metrics module and document instrumentation
