+++ src/Metallic/generation.rs
@@ -1,652 +1,20 @@
 use super::{Context, MetalError, Tensor};
-use crate::app_event::{AppEvent, LatencyRow, MemoryRow};
-use crate::metallic::instrumentation::{new_latency_collector, new_memory_collector, BlockMemorySnapshot, MemoryEvent, MemoryUsage};
+use crate::app_event::AppEvent;
+use crate::metallic::instrumentation::{new_latency_collector, new_memory_collector, MemoryEvent, MemoryUsage};
+use crate::metallic::metrics::{
+    build_latency_rows, build_memory_rows, build_model_memory_tree, log_interval_from_env, BlockStat, MemoryBlockStat, MemoryScopeStat,
+    MetricsLoggers, ProcessMemoryTracker, RollingStat, ScalarStat,
+};
 use crate::metallic::models::qwen25::Qwen25;
 use crate::metallic::Tokenizer;
-use chrono::{SecondsFormat, Utc};
 use rand::prelude::*;
-use serde::Serialize;
-use serde_json::json;
 use std::{
-    env,
-    fs::OpenOptions,
-    io::{self, BufWriter, Write},
     sync::mpsc,
     time::{Duration, Instant},
 };
-use sysinfo::{get_current_pid, Pid, ProcessRefreshKind, ProcessesToUpdate, System};
 
 const IM_START: &str = "<|im_start|>";
 const IM_END: &str = "<|im_end|>";
-
-#[derive(Clone)]
-struct ModelMemoryNode {
-    label: String,
-    bytes: usize,
-    children: Vec<ModelMemoryNode>,
-}
-
-impl ModelMemoryNode {
-    fn leaf(label: impl Into<String>, bytes: usize) -> Self {
-        Self {
-            label: label.into(),
-            bytes,
-            children: Vec::new(),
-        }
-    }
-
-    fn branch(label: impl Into<String>, children: Vec<ModelMemoryNode>) -> Self {
-        Self {
-            label: label.into(),
-            bytes: 0,
-            children,
-        }
-    }
-
-    fn total_bytes(&self) -> usize {
-        self.bytes + self.children.iter().map(|child| child.total_bytes()).sum::<usize>()
-    }
-}
-
-#[derive(Clone, Copy, Default)]
-struct RollingStat {
-    last: Duration,
-    total: Duration,
-    count: u64,
-}
-
-impl RollingStat {
-    fn record(&mut self, duration: Duration) {
-        self.last = duration;
-        self.total += duration;
-        self.count += 1;
-    }
-
-    fn last_ms(&self) -> f64 {
-        self.last.as_secs_f64() * 1000.0
-    }
-
-    fn average_ms(&self) -> f64 {
-        if self.count == 0 {
-            0.0
-        } else {
-            (self.total.as_secs_f64() * 1000.0) / self.count as f64
-        }
-    }
-}
-
-#[derive(Clone, Default)]
-struct BlockPhaseStat {
-    label: String,
-    stat: RollingStat,
-}
-
-#[derive(Clone, Default)]
-struct BlockStat {
-    total: RollingStat,
-    phases: Vec<BlockPhaseStat>,
-}
-
-impl BlockStat {
-    fn record_total(&mut self, duration: Duration) {
-        self.total.record(duration);
-    }
-
-    fn record_phase(&mut self, label: &str, duration: Duration) {
-        if let Some(entry) = self.phases.iter_mut().find(|entry| entry.label == label) {
-            entry.stat.record(duration);
-        } else {
-            let mut stat = RollingStat::default();
-            stat.record(duration);
-            self.phases.push(BlockPhaseStat {
-                label: label.to_string(),
-                stat,
-            });
-        }
-    }
-}
-
-#[derive(Clone, Default)]
-struct MemoryScopeStat {
-    current_pool_mb: f64,
-    peak_pool_mb: f64,
-    current_kv_mb: f64,
-    peak_kv_mb: f64,
-    current_kv_cache_mb: f64,
-    peak_kv_cache_mb: f64,
-}
-
-impl MemoryScopeStat {
-    fn update(
-        &mut self,
-        current_pool_bytes: usize,
-        current_kv_bytes: usize,
-        current_kv_cache_bytes: usize,
-        peak_pool_bytes: usize,
-        peak_kv_bytes: usize,
-        peak_kv_cache_bytes: usize,
-    ) {
-        self.current_pool_mb = bytes_to_mb(current_pool_bytes);
-        self.current_kv_mb = bytes_to_mb(current_kv_bytes);
-        self.current_kv_cache_mb = bytes_to_mb(current_kv_cache_bytes);
-        self.peak_pool_mb = self.peak_pool_mb.max(bytes_to_mb(peak_pool_bytes));
-        self.peak_kv_mb = self.peak_kv_mb.max(bytes_to_mb(peak_kv_bytes));
-        self.peak_kv_cache_mb = self.peak_kv_cache_mb.max(bytes_to_mb(peak_kv_cache_bytes));
-    }
-
-    fn has_data(&self) -> bool {
-        self.current_total_mb() > 0.0 || self.peak_total_mb() > 0.0
-    }
-
-    fn current_total_mb(&self) -> f64 {
-        self.current_pool_mb + self.current_kv_mb + self.current_kv_cache_mb
-    }
-
-    fn peak_total_mb(&self) -> f64 {
-        self.peak_pool_mb + self.peak_kv_mb + self.peak_kv_cache_mb
-    }
-
-    fn to_row(&self, label: String, level: u8) -> MemoryRow {
-        MemoryRow {
-            label,
-            level,
-            current_total_mb: self.current_total_mb(),
-            peak_total_mb: self.peak_total_mb(),
-            current_pool_mb: self.current_pool_mb,
-            peak_pool_mb: self.peak_pool_mb,
-            current_kv_mb: self.current_kv_mb,
-            peak_kv_mb: self.peak_kv_mb,
-            current_kv_cache_mb: self.current_kv_cache_mb,
-            peak_kv_cache_mb: self.peak_kv_cache_mb,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        }
-    }
-}
-
-#[derive(Clone, Default)]
-struct MemoryPhaseStat {
-    label: String,
-    scope: MemoryScopeStat,
-}
-
-#[derive(Clone, Default)]
-struct MemoryBlockStat {
-    scope: MemoryScopeStat,
-    phases: Vec<MemoryPhaseStat>,
-}
-
-impl MemoryBlockStat {
-    fn update_from_snapshot(&mut self, snapshot: &BlockMemorySnapshot) {
-        self.scope.update(
-            snapshot.current_pool_delta,
-            snapshot.current_kv_delta,
-            snapshot.current_kv_cache_delta,
-            snapshot.peak_pool_delta,
-            snapshot.peak_kv_delta,
-            snapshot.peak_kv_cache_delta,
-        );
-
-        for phase_snapshot in &snapshot.phases {
-            if let Some(phase_stat) = self.phases.iter_mut().find(|phase| phase.label == phase_snapshot.label) {
-                phase_stat.scope.update(
-                    phase_snapshot.current_pool_delta,
-                    phase_snapshot.current_kv_delta,
-                    phase_snapshot.current_kv_cache_delta,
-                    phase_snapshot.peak_pool_delta,
-                    phase_snapshot.peak_kv_delta,
-                    phase_snapshot.peak_kv_cache_delta,
-                );
-            } else {
-                let mut scope = MemoryScopeStat::default();
-                scope.update(
-                    phase_snapshot.current_pool_delta,
-                    phase_snapshot.current_kv_delta,
-                    phase_snapshot.current_kv_cache_delta,
-                    phase_snapshot.peak_pool_delta,
-                    phase_snapshot.peak_kv_delta,
-                    phase_snapshot.peak_kv_cache_delta,
-                );
-                self.phases.push(MemoryPhaseStat {
-                    label: phase_snapshot.label.clone(),
-                    scope,
-                });
-            }
-        }
-
-        self.phases
-            .retain(|phase| snapshot.phases.iter().any(|snap| snap.label == phase.label));
-    }
-}
-
-#[derive(Clone, Copy, Default)]
-struct ScalarStat {
-    baseline: Option<f64>,
-    current: f64,
-    peak: f64,
-}
-
-impl ScalarStat {
-    fn record(&mut self, value: f64) {
-        if self.baseline.is_none() {
-            self.baseline = Some(value);
-        }
-        self.current = value;
-        self.peak = self.peak.max(value);
-    }
-
-    fn baseline_mb(&self) -> Option<f64> {
-        self.baseline
-    }
-
-    fn to_row(&self, label: &str) -> MemoryRow {
-        MemoryRow {
-            label: label.to_string(),
-            level: 0,
-            current_total_mb: self.current,
-            peak_total_mb: self.peak,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        }
-    }
-}
-
-struct ProcessMemoryTracker {
-    system: System,
-    pid: Pid,
-}
-
-impl ProcessMemoryTracker {
-    fn new() -> Option<Self> {
-        let pid = get_current_pid().ok()?;
-        let mut system = System::new();
-        let targets = [pid];
-        if system.refresh_processes_specifics(
-            ProcessesToUpdate::Some(&targets),
-            false,
-            ProcessRefreshKind::nothing().with_memory(),
-        ) == 0
-        {
-            system.refresh_processes(ProcessesToUpdate::All, false);
-        }
-        Some(Self { system, pid })
-    }
-
-    fn sample_mb(&mut self) -> Option<f64> {
-        let pid = self.pid;
-        let targets = [pid];
-        if self.system.refresh_processes_specifics(
-            ProcessesToUpdate::Some(&targets),
-            false,
-            ProcessRefreshKind::nothing().with_memory(),
-        ) == 0
-        {
-            self.system.refresh_processes(ProcessesToUpdate::All, false);
-        }
-        self.system.process(pid).map(|process| process.memory() as f64 / 1024.0 / 1024.0)
-    }
-}
-
-struct MetricsLoggers {
-    memory: Option<JsonlLogger>,
-    latency: Option<JsonlLogger>,
-}
-
-impl MetricsLoggers {
-    fn new(interval: Duration) -> Self {
-        let timestamp = Utc::now().format("%Y%m%d%H%M%S").to_string();
-        let memory = JsonlLogger::create(&format!("{}-memory.jsonl", timestamp), interval).ok();
-        let latency = JsonlLogger::create(&format!("{}-latency.jsonl", timestamp), interval).ok();
-        let mut loggers = Self { memory, latency };
-        let now = Instant::now();
-        loggers.log_memory(&[], now, true);
-        if let Some(logger) = loggers.memory.as_mut() {
-            logger.last_logged = None;
-        }
-        loggers.log_latency(&[], now, true);
-        if let Some(logger) = loggers.latency.as_mut() {
-            logger.last_logged = None;
-        }
-        loggers
-    }
-
-    fn log_memory(&mut self, rows: &[MemoryRow], now: Instant, force: bool) {
-        if let Some(logger) = self.memory.as_mut() {
-            if let Err(err) = logger.log(rows, now, force) {
-                eprintln!("Failed to log memory metrics: {err}");
-                self.memory = None;
-            }
-        }
-    }
-
-    fn log_latency(&mut self, rows: &[LatencyRow], now: Instant, force: bool) {
-        if let Some(logger) = self.latency.as_mut() {
-            if let Err(err) = logger.log(rows, now, force) {
-                eprintln!("Failed to log latency metrics: {err}");
-                self.latency = None;
-            }
-        }
-    }
-}
-
-struct JsonlLogger {
-    writer: BufWriter<std::fs::File>,
-    interval: Duration,
-    last_logged: Option<Instant>,
-}
-
-impl JsonlLogger {
-    fn create(path: &str, interval: Duration) -> io::Result<Self> {
-        let file = OpenOptions::new().create(true).append(true).open(path)?;
-        Ok(Self {
-            writer: BufWriter::new(file),
-            interval,
-            last_logged: None,
-        })
-    }
-
-    fn log<T>(&mut self, rows: &[T], now: Instant, force: bool) -> io::Result<()>
-    where
-        T: Serialize,
-    {
-        let should_log = force
-            || self
-                .last_logged
-                .map(|last| now.duration_since(last) >= self.interval)
-                .unwrap_or(true);
-
-        if !should_log {
-            return Ok(());
-        }
-
-        let entry = json!({
-            "timestamp": Utc::now().to_rfc3339_opts(SecondsFormat::Millis, true),
-            "rows": rows,
-        });
-        serde_json::to_writer(&mut self.writer, &entry)?;
-        self.writer.write_all(b"\n")?;
-        self.writer.flush()?;
-        self.last_logged = Some(now);
-        Ok(())
-    }
-}
-
-fn bytes_to_mb(bytes: usize) -> f64 {
-    bytes as f64 / 1024.0 / 1024.0
-}
-
-fn usage_tracked_mb(usage: &MemoryUsage) -> f64 {
-    let pool_mb = bytes_to_mb(usage.pool_used.max(usage.pool_capacity));
-    let kv_mb = bytes_to_mb(usage.kv_used.max(usage.kv_capacity));
-    let kv_cache_mb = bytes_to_mb(usage.kv_cache_bytes);
-    pool_mb + kv_mb + kv_cache_mb
-}
-
-fn build_model_memory_tree(model: &Qwen25) -> ModelMemoryNode {
-    let mut children = Vec::new();
-
-    children.push(ModelMemoryNode::leaf("Token Embeddings", model.embed_weight.size_bytes()));
-    children.push(ModelMemoryNode::leaf("Output Projection", model.output_weight.size_bytes()));
-    children.push(ModelMemoryNode::leaf("Final Layer Norm", model.final_norm_gamma.size_bytes()));
-    children.push(ModelMemoryNode::leaf(
-        "RoPE Cache",
-        model.rope_cos_cache.size_bytes() + model.rope_sin_cache.size_bytes(),
-    ));
-
-    let block_nodes: Vec<ModelMemoryNode> = model
-        .blocks
-        .iter()
-        .enumerate()
-        .map(|(idx, block)| {
-            let attn_projections = ModelMemoryNode::branch(
-                "Attention Projections",
-                vec![
-                    ModelMemoryNode::leaf("Q weight", block.attn_q_weight.size_bytes()),
-                    ModelMemoryNode::leaf("K weight", block.attn_k_weight.size_bytes()),
-                    ModelMemoryNode::leaf("V weight", block.attn_v_weight.size_bytes()),
-                    ModelMemoryNode::leaf("Output weight", block.attn_out_weight.size_bytes()),
-                ],
-            );
-
-            let attn_biases = ModelMemoryNode::branch(
-                "Attention Biases",
-                vec![
-                    ModelMemoryNode::leaf("Q bias", block.attn_q_bias.size_bytes()),
-                    ModelMemoryNode::leaf("K bias", block.attn_k_bias.size_bytes()),
-                    ModelMemoryNode::leaf("V bias", block.attn_v_bias.size_bytes()),
-                ],
-            );
-
-            let feedforward = ModelMemoryNode::branch(
-                "Feedforward Projections",
-                vec![
-                    ModelMemoryNode::leaf("Gate weight", block.ffn_gate.size_bytes()),
-                    ModelMemoryNode::leaf("Up weight", block.ffn_up.size_bytes()),
-                    ModelMemoryNode::leaf("Down weight", block.ffn_down.size_bytes()),
-                ],
-            );
-
-            let feedforward_biases = ModelMemoryNode::branch(
-                "Feedforward Biases",
-                vec![
-                    ModelMemoryNode::leaf("Gate bias", block.ffn_gate_bias.size_bytes()),
-                    ModelMemoryNode::leaf("Up bias", block.ffn_up_bias.size_bytes()),
-                    ModelMemoryNode::leaf("Down bias", block.ffn_down_bias.size_bytes()),
-                ],
-            );
-
-            let norms = ModelMemoryNode::branch(
-                "Norm Parameters",
-                vec![
-                    ModelMemoryNode::leaf("Attention norm", block.attn_norm_gamma.size_bytes()),
-                    ModelMemoryNode::leaf("FFN norm", block.ffn_norm_gamma.size_bytes()),
-                ],
-            );
-
-            ModelMemoryNode::branch(
-                format!("Block {}", idx + 1),
-                vec![attn_projections, attn_biases, feedforward, feedforward_biases, norms],
-            )
-        })
-        .collect();
-
-    if !block_nodes.is_empty() {
-        children.push(ModelMemoryNode::branch("Transformer Blocks", block_nodes));
-    }
-
-    ModelMemoryNode::branch("Model Weights", children)
-}
-
-fn append_model_memory_rows(node: &ModelMemoryNode, level: u8, rows: &mut Vec<MemoryRow>) {
-    let total_mb = bytes_to_mb(node.total_bytes());
-    if total_mb <= 0.0 {
-        return;
-    }
-
-    rows.push(MemoryRow {
-        label: node.label.clone(),
-        level,
-        current_total_mb: total_mb,
-        peak_total_mb: total_mb,
-        current_pool_mb: 0.0,
-        peak_pool_mb: 0.0,
-        current_kv_mb: 0.0,
-        peak_kv_mb: 0.0,
-        current_kv_cache_mb: 0.0,
-        peak_kv_cache_mb: 0.0,
-        absolute_pool_mb: 0.0,
-        absolute_kv_mb: 0.0,
-        absolute_kv_cache_mb: 0.0,
-        show_absolute: false,
-    });
-
-    for child in &node.children {
-        append_model_memory_rows(child, level + 1, rows);
-    }
-}
-
-fn unattributed_host_row(host: &ScalarStat, tracked_mb: f64) -> Option<MemoryRow> {
-    if host.current <= 0.0 {
-        return None;
-    }
-
-    let current_gap = (host.current - tracked_mb).max(0.0);
-    let peak_gap = (host.peak - tracked_mb).max(current_gap);
-    if current_gap < 1.0 && peak_gap < 1.0 {
-        return None;
-    }
-
-    Some(MemoryRow {
-        label: "Unattributed Host Memory".to_string(),
-        level: 1,
-        current_total_mb: current_gap,
-        peak_total_mb: peak_gap.max(current_gap),
-        current_pool_mb: 0.0,
-        peak_pool_mb: 0.0,
-        current_kv_mb: 0.0,
-        peak_kv_mb: 0.0,
-        current_kv_cache_mb: 0.0,
-        peak_kv_cache_mb: 0.0,
-        absolute_pool_mb: 0.0,
-        absolute_kv_mb: 0.0,
-        absolute_kv_cache_mb: 0.0,
-        show_absolute: false,
-    })
-}
-
-fn baseline_host_row(host: &ScalarStat, tracked_mb: f64) -> Option<(MemoryRow, f64)> {
-    let baseline = host.baseline_mb()?;
-    if baseline <= 0.0 {
-        return None;
-    }
-
-    let gap = (baseline - tracked_mb).max(0.0);
-    if gap < 1.0 {
-        return None;
-    }
-
-    let row = MemoryRow {
-        label: "Process Baseline (Other)".to_string(),
-        level: 1,
-        current_total_mb: gap,
-        peak_total_mb: gap,
-        current_pool_mb: 0.0,
-        peak_pool_mb: 0.0,
-        current_kv_mb: 0.0,
-        peak_kv_mb: 0.0,
-        current_kv_cache_mb: 0.0,
-        peak_kv_cache_mb: 0.0,
-        absolute_pool_mb: 0.0,
-        absolute_kv_mb: 0.0,
-        absolute_kv_cache_mb: 0.0,
-        show_absolute: false,
-    };
-
-    Some((row, gap))
-}
-
-fn append_reserved_pool_rows(usage: &MemoryUsage, rows: &mut Vec<MemoryRow>) {
-    if usage.pool_capacity > 0 {
-        let capacity_mb = bytes_to_mb(usage.pool_capacity);
-        let used_mb = bytes_to_mb(usage.pool_used);
-        rows.push(MemoryRow {
-            label: "Tensor Pool (Reserved)".to_string(),
-            level: 1,
-            current_total_mb: capacity_mb,
-            peak_total_mb: capacity_mb,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        });
-        rows.push(MemoryRow {
-            label: "pool".to_string(),
-            level: 2,
-            current_total_mb: used_mb,
-            peak_total_mb: used_mb,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        });
-    }
-
-    if usage.kv_capacity > 0 {
-        let capacity_mb = bytes_to_mb(usage.kv_capacity);
-        let used_mb = bytes_to_mb(usage.kv_used);
-        let cache_mb = bytes_to_mb(usage.kv_cache_bytes);
-        rows.push(MemoryRow {
-            label: "KV Pool (Reserved)".to_string(),
-            level: 1,
-            current_total_mb: capacity_mb,
-            peak_total_mb: capacity_mb,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        });
-        rows.push(MemoryRow {
-            label: "kv".to_string(),
-            level: 2,
-            current_total_mb: used_mb,
-            peak_total_mb: used_mb,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        });
-        rows.push(MemoryRow {
-            label: "kv-cache".to_string(),
-            level: 2,
-            current_total_mb: cache_mb,
-            peak_total_mb: cache_mb,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        });
-    }
-}
 
 /// Generation configuration (defaults chosen by user)
 pub struct GenerationConfig {
@@ -888,13 +256,8 @@ where
     let kv_head_dim = kv_dim / n_kv_heads;
     let batch_size = 1; // Assuming batch size of 1 for now
 
-    let log_interval = env::var("METRICS_LOG_INTERVAL_SECS")
-        .ok()
-        .and_then(|value| value.parse::<u64>().ok())
-        .filter(|&secs| secs > 0)
-        .map(Duration::from_secs)
-        .unwrap_or_else(|| Duration::from_secs(5));
-    let mut metrics_loggers = MetricsLoggers::new(log_interval);
+    let log_interval = log_interval_from_env();
+    let mut metrics_loggers = MetricsLoggers::from_env(log_interval);
 
     let mut embed_stats = RollingStat::default();
     let mut forward_stats = RollingStat::default();
@@ -1084,8 +447,10 @@ where
 
         if latencies_ready {
             let rows = build_latency_rows(&embed_stats, &forward_stats, &block_stats, &output_stats, &sample_stats);
-            let log_now = Instant::now();
-            metrics_loggers.log_latency(&rows, log_now, false);
+            if let Some(loggers) = metrics_loggers.as_mut() {
+                let log_now = Instant::now();
+                loggers.log_latency(&rows, log_now, false);
+            }
             if ui_connected {
                 if tx.send(AppEvent::LatencyUpdate(rows)).is_err() {
                     ui_connected = false;
@@ -1104,8 +469,10 @@ where
                 &memory_output,
                 host_overheads,
             );
-            let log_now = Instant::now();
-            metrics_loggers.log_memory(&rows, log_now, false);
+            if let Some(loggers) = metrics_loggers.as_mut() {
+                let log_now = Instant::now();
+                loggers.log_memory(&rows, log_now, false);
+            }
             if ui_connected {
                 if tx.send(AppEvent::MemoryUpdate(rows)).is_err() {
                     ui_connected = false;
@@ -1123,148 +490,4 @@ where
     }
 
     Ok(())
-}
-
-fn build_latency_rows(
-    embed: &RollingStat,
-    forward: &RollingStat,
-    blocks: &[BlockStat],
-    output: &RollingStat,
-    sample: &RollingStat,
-) -> Vec<LatencyRow> {
-    let mut rows = Vec::new();
-
-    rows.push(LatencyRow {
-        label: "Embedding".to_string(),
-        last_ms: embed.last_ms(),
-        average_ms: embed.average_ms(),
-        level: 0,
-    });
-
-    rows.push(LatencyRow {
-        label: "Forward Step".to_string(),
-        last_ms: forward.last_ms(),
-        average_ms: forward.average_ms(),
-        level: 0,
-    });
-
-    for (idx, stat) in blocks.iter().enumerate() {
-        rows.push(LatencyRow {
-            label: format!("Block {}", idx + 1),
-            last_ms: stat.total.last_ms(),
-            average_ms: stat.total.average_ms(),
-            level: 1,
-        });
-        for phase in &stat.phases {
-            rows.push(LatencyRow {
-                label: phase.label.clone(),
-                last_ms: phase.stat.last_ms(),
-                average_ms: phase.stat.average_ms(),
-                level: 2,
-            });
-        }
-    }
-
-    rows.push(LatencyRow {
-        label: "Output".to_string(),
-        last_ms: output.last_ms(),
-        average_ms: output.average_ms(),
-        level: 0,
-    });
-
-    rows.push(LatencyRow {
-        label: "Sampling".to_string(),
-        last_ms: sample.last_ms(),
-        average_ms: sample.average_ms(),
-        level: 0,
-    });
-
-    rows
-}
-
-fn build_memory_rows(
-    model_memory: &ModelMemoryNode,
-    host: &ScalarStat,
-    embed: &MemoryScopeStat,
-    forward: &MemoryScopeStat,
-    forward_usage: Option<MemoryUsage>,
-    blocks: &[MemoryBlockStat],
-    output: &MemoryScopeStat,
-    host_overheads: &[(String, usize)],
-) -> Vec<MemoryRow> {
-    let mut rows = Vec::new();
-
-    append_model_memory_rows(model_memory, 0, &mut rows);
-
-    let mut host_tracked_mb = bytes_to_mb(model_memory.total_bytes());
-
-    for (label, bytes) in host_overheads {
-        if *bytes == 0 {
-            continue;
-        }
-        let mb = bytes_to_mb(*bytes);
-        host_tracked_mb += mb;
-        rows.push(MemoryRow {
-            label: label.clone(),
-            level: 0,
-            current_total_mb: mb,
-            peak_total_mb: mb,
-            current_pool_mb: 0.0,
-            peak_pool_mb: 0.0,
-            current_kv_mb: 0.0,
-            peak_kv_mb: 0.0,
-            current_kv_cache_mb: 0.0,
-            peak_kv_cache_mb: 0.0,
-            absolute_pool_mb: 0.0,
-            absolute_kv_mb: 0.0,
-            absolute_kv_cache_mb: 0.0,
-            show_absolute: false,
-        });
-    }
-
-    if host.current > 0.0 || host.peak > 0.0 {
-        let mut explained_host_mb = host_tracked_mb;
-        rows.push(host.to_row("Host Memory (MB)"));
-        if let Some(usage) = forward_usage {
-            explained_host_mb += usage_tracked_mb(&usage);
-            append_reserved_pool_rows(&usage, &mut rows);
-        }
-
-        if let Some((baseline_row, baseline_mb)) = baseline_host_row(host, explained_host_mb) {
-            explained_host_mb += baseline_mb;
-            rows.push(baseline_row);
-        }
-
-        if let Some(unattributed) = unattributed_host_row(host, explained_host_mb) {
-            rows.push(unattributed);
-        }
-    }
-
-    if embed.has_data() {
-        rows.push(embed.to_row("Embedding".to_string(), 0));
-    }
-
-    if forward.has_data() || forward_usage.is_some() {
-        let forward_row = forward.to_row("Forward Step".to_string(), 0);
-        rows.push(forward_row);
-
-        for (idx, block_stat) in blocks.iter().enumerate() {
-            if !block_stat.scope.has_data() {
-                continue;
-            }
-            rows.push(block_stat.scope.to_row(format!("Block {}", idx + 1), 1));
-            for phase in &block_stat.phases {
-                if !phase.scope.has_data() {
-                    continue;
-                }
-                rows.push(phase.scope.to_row(phase.label.clone(), 2));
-            }
-        }
-    }
-
-    if output.has_data() {
-        rows.push(output.to_row("Output".to_string(), 0));
-    }
-
-    rows
 }
