+++ src/Metallic/mod.rs
@@ -12,6 +12,7 @@ pub mod context;
 pub mod encoder;
 pub mod error;
 pub mod instrumentation;
+pub mod metrics;
 pub mod operation;
 pub mod pool;
 pub mod resource_cache;
